Call plug#AfterConfig much earlier than VimEnter
For long time I was making calls this way to ensure that `g:loaded_*`
was properly populated not to call functions for deleted plugins
Well this proved to be problematic for lua plugins as functionality
changed. Specifically I wouldn't get `completion_nvim` to work if nvim
was started with an argument. This was fixed by setting up just after
Plug calls
I addressed the `g:loaded_` checks by checking if the functions exist
instead. Overall is a much better way to setup