Major improvement to toggle lunch tool
Made it more customizable search windows in i3. Ex: [class=X]
Moved argument 2, command to execute if app does not exist to be a
member of xdotool. This so that i3 config can be less verbose
Also allows for more customizable. i.e: spotify
Also added custom sized for scratch, and ranger for surbook