fix(packer): packer_compiled.vim was being sourced twice
This was in reality because I was placing the file under `plugins`. So
vim was sourcing it, and packer as well.
In reality there is an issue where if you move it then nobody seems to
source it then.
Resolution, don't move it! If you get this issue again, just make sure
you don't have multiple packer_compiled.vim files