fix: removing alt combined with enter
It's very difficult to get the timing of the key right.
It was being annoying because I was having to wait sometimes to press
the alt key

The main thing for the dual is for when your hands are in the mouse and
you get a pop up. It's convenient to have an enter key on the other
hand.

So, I am experimenting with leaving alt fix there and enter on the right
of the windows key